+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 -   The contour functions gain a `clip` argument to only show contours in an area defined by a polygon.
+-   The `kriging` argument of the contour functions now can be a numeric to control de number of pixels used.
 
 # metR 0.14.1